requirements.txt
flask==2.0.1
werkzeug==2.0.3
torch>=1.7.0
torchvision>=0.8.1
numpy>=1.18.5
opencv-python>=4.1.2
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
tqdm>=4.41.0
ultralytics
opencv-python-headless
Pillow
torch
pandas
seaborn
matplotlib
gitpython>=3.1.30